requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
bleach==6.0.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
discord.py==2.0.1
docopt==0.6.2
docutils==0.19
dotty-dict==1.3.1
frozenlist==1.3.1
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.13.0
google-auth-oauthlib==0.5.3
gotrue==1.0.0
gspread==5.6.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
invoke==1.7.3
jaraco.classes==3.2.3
keyring==23.13.1
more-itertools==9.0.0
multidict==6.0.2
oauthlib==3.2.2
packaging==23.0
pipreqs==0.4.11
pkginfo==1.9.6
postgrest-py==0.10.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-gitlab==3.13.0
python-semantic-release==7.33.0
pywin32-ctypes==0.2.0
readme-renderer==37.3
realtime==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
storage3==0.5.0
supabase==1.0.0
supafunc==0.2.2
tomlkit==0.11.6
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
websockets==10.4
yarg==0.1.9
yarl==1.8.1
zipp==3.12.1
